feat(structure,rbp): widen saved-feature outputs to 64-bit to match dtype map
populate_structure.py:
  - sense_start, sense_start_from_end: int32 -> int64
  - sense_start_norm, sense_start_from_end_norm: float32 -> float64
  - sense_dist_to_{canonical,closest}_{stop,start}: float32 -> float64
  - sense_mrna_dist_to_{canonical,closest}_stop: float32 -> float64
  - Drop the .astype(np.float32) calls that fed those arrays (the destinations
    are now float64, so the explicit downcasts were pure overhead).

populate_rbp.py:
  - RBP_* affinity scores: float32 -> float64.

Local intermediates feeding numba kernels (seq_indices, seq_ints, background
probs) are kept narrow on purpose -- they don't become saved feature columns.

+++ src/tauso/populate/populate_rbp.py
@@ -88,7 +88,7 @@ def process_rbp(task, sequences, background_probs_arr):
     col_name = task["col_name"]
     n_rows = len(sequences)
 
-    scores = np.zeros(n_rows, dtype=np.float32)
+    scores = np.zeros(n_rows, dtype=np.float64)
     for matrix in matrices:
         for i in range(n_rows):
             scores[i] += calculate_total_affinity_numba(sequences[i], matrix, background_probs_arr[i])
